feat(math)!: move `useClamp` to `@vueuse/math` (#1810)
