+++ meta/packages.ts
@@ -32,9 +32,7 @@ export const packages: PackageManifest[] = [
     name: 'math',
     display: 'Math',
     description: 'Math functions for VueUse',
-    author: 'Jacob Clevenger<https://github.com/wheatjs>',
     external: [
-      '@vueuse/core',
       '@vueuse/shared',
     ],
   },
