+++ packages/core/index.ts
@@ -17,7 +17,6 @@ export * from './useBreakpoints'
 export * from './useBroadcastChannel'
 export * from './useBrowserLocation'
 export * from './useCached'
-export * from './useClamp'
 export * from './useClipboard'
 export * from './useColorMode'
 export * from './useConfirmDialog'
