+++ packages/core/useOffsetPagination/index.ts
@@ -2,7 +2,7 @@ import type { ComputedRef, Ref, UnwrapNestedRefs } from 'vue-demi'
 import { computed, isRef, reactive, unref, watch } from 'vue-demi'
 import { noop, syncRef } from '@vueuse/shared'
 import type { MaybeRef } from '@vueuse/shared'
-import { useClamp } from '../useClamp'
+import { useClamp } from '../../math/useClamp'
 
 export interface UseOffsetPaginationOptions {
   /**
@@ -62,7 +62,7 @@ export function useOffsetPagination(options: UseOffsetPaginationOptions): UseOff
     onPageCountChange = noop,
   } = options
 
-  const currentPageSize = useClamp(pageSize, 1, Infinity)
+  const currentPageSize = computed(() => Math.min(unref(pageSize), 1))
 
   const pageCount = computed(() => Math.ceil((unref(total)) / unref(currentPageSize)))
