+++ packages/math/index.ts
@@ -1,4 +1,5 @@
 export * from './logicAnd'
 export * from './logicNot'
 export * from './logicOr'
+export * from './useClamp'
 export * from './useProjection'
