+++ packages/math/package.json
@@ -1,8 +1,8 @@
 {
   "name": "@vueuse/math",
-  "version": "8.8.1",
+  "version": "8.9.1",
   "description": "Math functions for VueUse",
-  "author": "Jacob Clevenger<https://github.com/wheatjs>",
+  "author": "Anthony Fu <https://github.com/antfu>",
   "license": "MIT",
   "funding": "https://github.com/sponsors/antfu",
   "homepage": "https://github.com/vueuse/vueuse/tree/main/packages/math#readme",
