+++ packages/math/useClamp/index.md
@@ -1,5 +1,5 @@
 ---
-category: Utilities
+category: '@Math'
 ---
 
 # useClamp
@@ -9,7 +9,7 @@ Reactively clamp a value between two other values.
 ## Usage
 
 ```ts
-import { useClamp } from '@vueuse/core'
+import { useClamp } from '@vueuse/math'
 
 const min = ref(0)
 const max = ref(10)
